requirements.txt
aiofiles==23.2.1
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
cryptography==46.0.2
ecdsa==0.19.1
fastapi==0.104.1
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-texttospeech==2.31.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.75.1
gTTS==2.5.4
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.2
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pydub==0.25.1
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
rapidfuzz==3.5.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.0
websockets==15.0.1
